+++ admin/class-tmsm-aquos-spa-booking-admin.php
@@ -308,4 +308,322 @@ class Tmsm_Aquos_Spa_Booking_Admin {
 		);
 
 	}
+
+	/**
+	 * Regenerate Aquos Prices
+	 *
+	 * @param bool $echo
+	 */
+	function regenerate_aquos_prices($echo = false){
+
+		$output = '';
+
+		global $wpdb;
+
+
+		// Finding missings Aquos IDs
+		$missing_aquos_ids = [];
+		$products = wc_get_products(['numberposts' => -1, 'post_status' => 'publish']);
+		foreach($products as $product){
+
+			if ($product->get_type() == "variable") {
+				foreach ($product->get_available_variations() as $variation) {
+
+						$aquos_id = get_post_meta($variation['variation_id'], '_aquos_id', true);
+
+						if(empty($aquos_id)){
+							$missing_aquos_ids[$variation['variation_id']] = '<a href="'.get_edit_post_link($product->get_id()).'">'.$product->get_name(). '</a> / ' . ' - '.$variation['sku']. ' variation ID '. $product->get_id().' / aquos_id missing';
+						}
+
+				}
+			}
+			else{
+				$aquos_id = get_post_meta($product->get_id(), '_aquos_id', true);
+				if(empty($aquos_id)){
+					$missing_aquos_ids[$product->get_id()] = '<a href="'.get_edit_post_link($product->get_id()).'">'.$product->get_name(). '</a> / ' . ' product ID '.$product->get_id().' / aquos_id missing';
+				}
+			}
+
+		}
+		if(!empty($missing_aquos_ids)){
+			$output .= __( 'This section finds if Aquos IDs are missing for the web service synchronization:', 'tmsm-aquos-spa-booking' );
+			$output .= '<br>';
+			$output .= implode('<br>', $missing_aquos_ids);
+			$output .= '<br>';
+			$output .= '<br>';
+		}
+
+
+
+		// Finding non matching prices
+		$meta_key_aquos_price = '_aquos_price';
+		$non_matching_prices = [];
+
+		$wpdb->delete( $wpdb->postmeta, [ 'meta_key' => $meta_key_aquos_price ] );
+
+		$results = $wpdb->get_results("SELECT * FROM $wpdb->postmeta WHERE meta_key = '_aquos_id' AND meta_value != '' ");
+
+		if(!empty($results)){
+			$count_meta = count( $results );
+			$errors_nb = 0;
+
+			//echo sprintf(__( 'Parsing %d products', 'tmsm-aquos-spa-booking' ), $count_meta);
+
+			foreach ( $results as $result ) {
+
+				$non_matching_prices[$result->post_id] = '';
+				$non_matching_prices[$result->post_id] .= 'product id '.$result->post_id . ' / ';
+
+				$aquos_price = [];
+				$aquos_ids   = explode( '+', $result->meta_value );
+				foreach ( $aquos_ids as $aquos_id ) {
+					$price = self::get_product_price_from_aquos_or_product_id( $aquos_id, $result->post_id );
+					$aquos_price[] = $price;
+					if($price == 0){
+						$non_matching_prices[$result->post_id] .= sprintf(__( 'Price not found for product %s', 'tmsm-aquos-spa-booking' ), $result->post_id);
+						$non_matching_prices[$result->post_id] .= ' / ';
+					}
+				}
+				$result->aquos_price = join( '+', $aquos_price );
+				$non_matching_prices[$result->post_id] .= 'detailed price '.$result->aquos_price . ' / ';
+
+				// Checking if calculated price matches product price
+				$product = wc_get_product($result->post_id);
+				$product_price = 0;
+				if(!empty($product)){
+					$product_price = $product->get_price();
+				}
+				$calculated_price = 0;
+				foreach($aquos_price as $item_price){
+					$calculated_price+=$item_price;
+				}
+				$non_matching_prices[$result->post_id] .= 'calculated price '.$calculated_price . ' / ';
+
+				if($calculated_price != $product_price){
+					$non_matching_prices[$result->post_id] .= '<b>NON MATCHING PRICE with '.$product_price . '</b> / ';
+
+					if(!empty($product)){
+						$non_matching_prices[$result->post_id] .= 'product is '. $product->get_status() . ' / ';
+						if($product->get_status() === 'draft'){
+							unset($non_matching_prices[$result->post_id]);
+						}
+						else{
+							$parent = wc_get_product($product->get_parent_id());
+							if(!empty($parent)){
+								$non_matching_prices[$result->post_id] .= 'variation is '.  $parent->get_status(). ' / ';
+								$non_matching_prices[$result->post_id] = '<a href="'.get_edit_post_link($parent->get_id()).'">'.$parent->get_name(). '</a> / ' . $non_matching_prices[$result->post_id];
+								if($parent->get_status() === 'draft'){
+									unset($non_matching_prices[$result->post_id]);
+								}
+							}
+						}
+
+					}
+
+					if(!empty($product) && ! $product->get_status() == 'published'){
+						//echo 'product is draft '. $product->get_status() . ' / ';
+					}
+					if(!empty($product) && $product->is_type('variation') ){
+						$parent = wc_get_product($product->get_parent_id());
+						if(!empty($parent) && ! $parent->get_status() == 'published'){
+							//echo 'variation is draft / ';
+						}
+					}
+
+					$errors_nb ++;
+				}
+				else{
+					unset($non_matching_prices[$result->post_id]);
+				}
+
+				if( defined('TMSM_AQUOS_SPA_BOOKING_DEBUG') && TMSM_AQUOS_SPA_BOOKING_DEBUG ){
+					//print_r( $result );
+				}
+
+				// Insert the values
+				$wpdb->query($wpdb->prepare(" INSERT INTO $wpdb->postmeta ( post_id, meta_key, meta_value ) VALUES ( %d, %s, %s ) ",
+					$result->post_id,
+					$meta_key_aquos_price,
+					$result->aquos_price
+				)
+				);
+
+			}
+
+
+		}
+
+		if(!empty($non_matching_prices)){
+			$output .= __( 'This section lists the products/variations with a non-matching price for the web service synchronization:', 'tmsm-aquos-spa-booking' );
+			$output .= '<br>';
+			$output .= implode('<br>', $non_matching_prices);
+			$output .= '<br>';
+			$output .= '<br>';
+		}
+
+		// If any error occured, output is defined
+		if(!empty($output)){
+			if ( $echo ) {// print result
+				echo $output;
+			}
+			else {// email result
+
+				$blogname = esc_html( get_bloginfo( 'name' ) );
+				$email    = stripslashes( get_option( 'admin_email' ) );
+				$subject  = sprintf(__( '%s: TMSM Aquos Spa Booking regeneration error on %s', 'tmsm-aquos-spa-booking' ), $blogname, date('Y-m-d'));
+
+				$message = $output;
+
+				$headers = 'Auto-Submitted: auto-generated; Content-Type: text/html; charset=UTF-8;';
+				$email_sent = wp_mail( $email, $subject, $message, $headers );
+				if (  TMSM_AQUOS_SPA_BOOKING_DEBUG && !$email_sent) {
+					error_log('Error email sent');
+				}
+			}
+		}
+	}
+
+	/**
+	 * Get Price From Aquos or Product ID
+	 *
+	 * @param int $aquos_id
+	 * @param int $product_id
+	 *
+	 * @return int
+	 */
+	function get_product_price_from_aquos_or_product_id(int $aquos_id, int $product_id){
+
+		global $wpdb;
+
+		$price = 0;
+
+		if(is_multisite() && get_current_blog_id() === 6) { // Rennes
+			switch ( $aquos_id ) {
+				case 338: // Parcours Aquatonic
+					$price = 20;
+					break;
+				case 191: // Modelage nuque et cuir chevelu
+					$price = 18;
+					break;
+				case 538: // Modelage dos d'accueil
+					$price = 18;
+					break;
+				case 427: // Une nuit Oceania
+					$price = 80;
+					break;
+			}
+		}
+		if(is_multisite() && get_current_blog_id() === 8) { // Nantes
+			switch ( $aquos_id ) {
+				case 338: // Parcours Aquatonic
+					$price = 23;
+					break;
+				case 191: // Modelage nuque et cuir chevelu
+					$price = 18;
+					break;
+				case 538: // Modelage dos d'accueil
+					$price = 18;
+					break;
+			}
+		}
+		if ( is_multisite() && get_current_blog_id() === 9 ) { // Paris
+			switch ( $aquos_id ) {
+				case 368: // Parcours Aquatonic (Paris)
+					$price = 29;
+					break;
+				case 470: // Accès Espace Bien-être en complément d'un soin
+					$price = 30;
+					break;
+				case 495: // Balnéo privative en duo
+					$price = 39;
+					break;
+				case 847: // Supplément French Manucure
+					$price = 5;
+					break;
+			}
+		}
+
+		if( $price == 0 ){
+			//echo 'findind product with aquos id '.$aquos_id.' / ';
+
+			$results = $wpdb->get_results(
+				$wpdb->prepare(
+					"SELECT * FROM $wpdb->postmeta WHERE meta_key = '_aquos_id' AND meta_value = '%s'", $aquos_id
+				)
+			);
+			//echo "SELECT * FROM $wpdb->postmeta WHERE meta_key = '_aquos_price' AND meta_value = " .$aquos_id . ' / ';
+
+			// Variable price procucts (like e-check)
+			if ( count( $results ) > 4 ) {
+				$product = wc_get_product( $product_id );
+				$price   = $product->get_price();
+				//echo 'product price variable '.$price . ' / ';
+			}
+
+			// Other cases
+			else{
+
+				$result = null;
+
+				// Take only the first result
+				if ( is_array( $results ) && count( $results ) > 0 ) {
+					$result = $results[0];
+				}
+
+				//print_r($results);
+				if(!empty($result)){
+					//echo 'finding product with post_id '.$result->post_id . ' / ';
+					$product = wc_get_product($result->post_id);
+
+					if(!empty($product)){
+
+						$price = $product->get_price();
+						//echo 'product price '.$price . ' / ';
+
+						/*$parent_id = $product->get_parent_id();
+						if(!empty($parent_id)){
+							echo 'found parent product '.$parent_id. ' / ';
+							$parent_product = wc_get_product($parent_id);
+							if(!empty($parent_product)){
+								$price = $parent_product->get_price();
+								echo 'parent price '.$price. ' / ';
+							}
+						}*/
+
+
+					}
+				}
+			}
+
+
+
+
+			/*$product = wc_get_product($product_id);
+
+			if(!empty($product)){
+
+				$price = $product->get_price();
+				echo 'product price '.$price . ' / ';
+
+				$parent_id = $product->get_parent_id();
+				if(!empty($parent_id)){
+					echo 'found parent product '.$parent_id. ' / ';
+					$parent_product = wc_get_product($parent_id);
+					if(!empty($parent_product)){
+						$price = $parent_product->get_price();
+						echo 'parent price '.$price. ' / ';
+					}
+				}
+
+
+			}*/
+		}
+		else{
+			//echo 'hard coded price '.$price . ' for '.$aquos_id. ' / ';
+		}
+
+		return $price;
+	}
+
+
 }