+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Alert when Aquos ID field is empty in a variation
+* Cron action to regenerate prices and check price consistency and missing aquos ids, executed daily at midnight
 
 ### 1.4.0: December 15th, 2020
 * Tool to see if products miss the _aquos_id field